+++ tests/integration/test_ctgan.py
@@ -12,7 +12,7 @@ model are not checked.
 import numpy as np
 import pandas as pd
 
-from ctgan.synthesizers import CTGANSynthesizer
+from ctgan.synthesizers.ctgan import CTGANSynthesizer
 
 
 def test_ctgan_dataframe():
